requirements.txt
asgiref==3.5.2
autopep8==1.7.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.13.0
django-deep-serializer==0.1.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
Faker==15.3.2
filelock==3.7.1
greenlet==1.1.3
idna==3.3
lxml==4.9.1
mypy==0.991
mypy-extensions==0.4.3
mysqlclient==2.1.1
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.9.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.14.1
virtualenvwrapper-win==1.2.7
